refactor: extract mobility tab to views/MobilityView.jsx
Strangler-fig extraction #3 of the sprint (#76, part of #52). Moves the
~320-line mobility tab out of the monolith with no behaviour change.

- new views/MobilityView.jsx owns its local mobOpen state and renders the
  routines library + tracking log; takes no props
- MOBILITY_ROUTINES + MOBILITY_STREAK_NOTE (pure data) move to the new
  constants/mobility.js (excluded from coverage, per the constants policy);
  mobilityLog is imported directly from constants/logs.js
- monolith renders {view === 'mobility' && <MobilityView />} and drops the
  now-unused MOBILITY_ROUTINES / MOBILITY_STREAK_NOTE / mobOpen / mobilityLog
- render test covers the library, expand/collapse with rehab blocks, and the
  tracking log; per-file coverage gated at 80/80/70 (measured 100 / 100 / 88.9)

Monolith: 6,303 -> 5,834 lines.
