+++ web/src/erg-dashboard.jsx
@@ -16,6 +16,7 @@ import ErgView from './views/ErgView.jsx';
 import JournalView from './views/JournalView.jsx';
 import RecoveryView from './views/RecoveryView.jsx';
 import StrengthView from './views/StrengthView.jsx';
+import MobilityView from './views/MobilityView.jsx';
 import WorkoutItem from './components/WorkoutItem.jsx';
 import LogSessionForm from './components/LogSessionForm.jsx';
 import LogEntry from './components/LogEntry.jsx';
@@ -75,7 +76,7 @@ import {
   autoregulate,
   calcReadiness,
 } from './utils/analysis.js';
-import { mobilityLog, nutritionLog, recoveryLog } from './constants/logs.js';
+import { nutritionLog, recoveryLog } from './constants/logs.js';
 import {
   ADAPTIVE_RULES,
   RULE_EVOLUTION,
@@ -394,158 +395,6 @@ const MOBILITY_WARMUP = [
 // Research: subjective measures often outperform objective for catching
 // overreaching. One number per session + daily feel. Logged with sessions.
 
-// ── MOBILITY / YOGA — an integral training pillar ─────────────
-// Not accessory afterthought. Given the left hamstring/glute rehab
-// (physio), mobility is load-bearing for staying healthy enough to
-// train. Three strands: soft-tissue (foam roll), pre-session drills
-// (prime the body before erg/lift), and yoga flows (longer reset).
-// MOBILITY_ROUTINES = the library (routines on hand). mobilityLog =
-// the tracking (consistency as a visible pillar).
-const MOBILITY_ROUTINES = [
-  {
-    id: 'foam',
-    name: 'Foam Roll / Soft Tissue',
-    icon: '🧻',
-    color: '#00d4ff',
-    when: 'Post-session or evening. Anytime tissue feels tight.',
-    why: 'Manages tissue quality + tightness around the rowing posture (thoracic, lats, hips) and the rehab area. Down-regulates after hard sessions.',
-    blocks: [
-      {
-        move: 'Thoracic spine — 60-90s',
-        visual:
-          'Lie on your back, roller across your mid-back (under the shoulder blades). Hands behind head, elbows in. Lift hips slightly and roll from mid-back up toward the shoulders — pause and breathe on tight spots. Can gently arch back over the roller at a sticky segment.',
-      },
-      {
-        move: 'Lats — 45s/side',
-        visual:
-          'Lie on one side, arm overhead, roller in the armpit/upper-ribs area. Roll the line from armpit down toward the lower ribs. Rotate slightly front-to-back to find the lat. This is the reach/finish muscle in the stroke.',
-      },
-      {
-        move: 'Glutes / piriformis — 60s/side',
-        rehab: true,
-        visual:
-          'Sit on the roller (or a ball for more pressure). Cross one ankle over the opposite knee (figure-4), lean slightly toward the crossed side. Roll the meat of the glute slowly. Ball into the deepest spot and breathe. Key rehab tissue — give the left side time.',
-      },
-      {
-        move: 'Hamstrings — 60s/side',
-        rehab: true,
-        visual:
-          "Sit with the roller under one thigh, hands behind for support, leg straight. Roll from just under the sit-bone (origin) down toward the back of the knee. Turn the foot in/out to hit different parts. Extra time and care on the left — back off if it's sharp near the origin (rehab area).",
-      },
-      {
-        move: 'Quads / IT band — 45-60s/side',
-        visual:
-          'Face-down (plank-ish) with the roller under the front of one thigh. Roll hip-to-knee. Rotate slightly to the outer thigh (ITB) — that side is often tender, keep pressure tolerable.',
-      },
-      {
-        move: 'Calves — 45s/side',
-        visual:
-          'Seated, roller under one calf, other leg crossed over for extra load. Hands lift the hips. Roll ankle-to-knee, rotate the foot in/out to cover inner/outer calf.',
-      },
-    ],
-    note: 'Slow. Breathe into tight spots. NOT a pain contest — discomfort ok, sharp pain no. Extra care + time on the left hamstring/glute.',
-  },
-  {
-    id: 'prep',
-    name: 'Pre-Session Mobility / Prime',
-    icon: '🔑',
-    color: '#34d399',
-    when: "Before EVERY erg + every lift (the 'mobility first' in session notes).",
-    why: 'Primes the body to move well before loading it. Cheap insurance — opens hips/t-spine/shoulders so the catch and the squat have range. Protects the rehab area by warming it before load.',
-    blocks: [
-      {
-        move: 'Cat-cow — 8-10 slow reps',
-        visual:
-          'On hands and knees (hands under shoulders, knees under hips). Inhale: drop the belly, lift the chest and tailbone (cow). Exhale: round the spine up, tuck the chin and tailbone (cat). Move slowly, segment by segment, led by the breath.',
-      },
-      {
-        move: 'Hip 90/90 transitions — 5/side',
-        visual:
-          'Sit on the floor, both legs bent at 90°: front shin across in front of you, back shin out to the side. Keeping chest tall, rotate both knees over to the other side so the legs swap front/back. Flow side to side. Opens internal/external hip rotation for the catch.',
-      },
-      {
-        move: "World's greatest stretch — 4/side",
-        visual:
-          'From a lunge (right foot forward), left hand on the floor inside the front foot. Drive the right elbow down toward the floor near the right foot, then rotate and reach that same arm up to the ceiling, opening the chest. Return and repeat. Full-body opener — hip, t-spine, hamstring.',
-      },
-      {
-        move: 'Leg swings, front-back + lateral — 10/side',
-        rehab: true,
-        visual:
-          'Hold a wall/rack for balance. Swing one leg forward and back like a pendulum, relaxed, increasing range gradually — keep the torso still. Then swing the same leg side-to-side across the body. Dynamic hamstring/glute prep — controlled, never forced (rehab warm-up).',
-      },
-      {
-        move: 'Thoracic open-books — 6/side',
-        visual:
-          'Lie on your side, knees bent and stacked, arms extended together in front at shoulder height. Keeping knees down, sweep the top arm up and over to the other side, opening the chest toward the ceiling, follow with your eyes. Return. Restores rotation lost to the rowing posture.',
-      },
-      {
-        move: 'Band shoulder pass-throughs — 10',
-        visual:
-          "Hold a band (or stick) wide in both hands in front of you. Keeping arms straight, raise it overhead and back behind you as far as comfortable, then return — a big arc. Widen your grip if it's tight. Opens the shoulders for the catch and overhead.",
-      },
-      {
-        move: 'Glute bridges — 12',
-        rehab: true,
-        visual:
-          "Lie on your back, knees bent, feet flat hip-width. Push through the heels and lift the hips until knees-hips-shoulders form a line — squeeze the glutes hard at the top, don't arch the low back. Lower slowly. Fires the glutes before squat/erg (rehab activation — don't skip).",
-      },
-    ],
-    note: "5-8min, dynamic not static (save long holds for post). The glute bridges + leg swings double as rehab activation — don't skip them.",
-  },
-  {
-    id: 'yoga',
-    name: 'Yoga Flow',
-    icon: '🧘',
-    color: '#a78bfa',
-    when: 'Recovery slots (Thu AM), rest days, evenings. The longer reset.',
-    why: 'Longer holds for genuine range + a parasympathetic down-shift (recovery, sleep, HRV). Counters the flexion-dominant rowing posture with extension + rotation.',
-    blocks: [
-      {
-        move: 'Down dog → forward fold — 1-2min',
-        visual:
-          'From hands and knees, tuck toes and lift hips up and back into an inverted-V (down dog) — heels reaching toward the floor, spine long, pedal the feet. Then walk hands back to the feet and hang into a forward fold, knees soft, head heavy. Settles the whole posterior chain.',
-      },
-      {
-        move: 'Low lunge + half-split — 1min/side',
-        rehab: true,
-        visual:
-          'Step into a low lunge (right foot forward, left knee down), sink the hips forward to stretch the back hip flexor. Then shift the hips back, straightening the front leg with toes up, folding gently over it (half-split) for the hamstring. Gentle on the left — physio-comfortable range only (rehab).',
-      },
-      {
-        move: 'Pigeon — 1-2min/side',
-        rehab: true,
-        visual:
-          'From down dog or hands-and-knees, bring the right shin forward and across, angled under your torso, left leg extended straight back. Square the hips and fold forward over the front shin, breathing into the deep glute. Ease into the left side slowly — no sharp stretch (rehab area).',
-      },
-      {
-        move: 'Thread-the-needle — 1min/side',
-        visual:
-          'On hands and knees, slide the right arm underneath the body and across to the left, shoulder and temple coming toward the floor, palm up. Rest there and breathe into the upper-back rotation. Return and switch. T-spine release.',
-      },
-      {
-        move: 'Cobra / sphinx — 1min',
-        visual:
-          'Lie face-down. Sphinx: prop on the forearms, elbows under shoulders, gently lifting the chest with a long low back. Cobra: hands by the ribs, press the chest higher, shoulders down away from the ears. Spinal extension — the direct antidote to rowing flexion.',
-      },
-      {
-        move: 'Supine twist — 1min/side',
-        visual:
-          'Lie on your back, hug both knees in, then let them drop to one side while you extend the opposite arm out and turn the head away. Keep both shoulders grounded. Breathe into the low-back and hip rotation. Switch sides.',
-      },
-      {
-        move: 'Legs-up-wall + breath — 3-5min',
-        visual:
-          'Sit side-on to a wall, then swing the legs up it as you lie back, hips close to or a few inches from the wall, arms relaxed. Just breathe slowly — long exhales. The down-regulation finish: shifts you into recovery (parasympathetic) mode.',
-      },
-    ],
-    note: 'Breath-led, never force. Pigeon + half-split touch the rehab area — back off to physio-comfortable range, no sharp stretch. Great on poor-sleep days (the down-shift helps).',
-  },
-];
-
-const MOBILITY_STREAK_NOTE =
-  "Consistency is the metric here, not intensity. The prep work especially — it's rehab activation disguised as a warm-up. Dropping it is the flag, not a quiet skip.";
-
 // ── ANNUAL MACRO ARC ──────────────────────────────────────────
 // TARGET: World Rowing Virtual Indoor Championships, ~late Feb 2027
 // (date provisional — confirm when World Rowing announces 2027).
@@ -787,7 +636,6 @@ export default function App() {
   const [ftp, setFtp] = useState(190);
   const [progTab, setProgTab] = useState('phases'); // phases | week | year
   const [nowTick, setNowTick] = useState(new Date()); // for date-awareness (day rollover)
-  const [mobOpen, setMobOpen] = useState(null); // which mobility routine is expanded
   const [vw, setVw] = useState(
     typeof window !== 'undefined' ? window.innerWidth : 1200
   ); // viewport width → responsive layout
@@ -5831,324 +5679,7 @@ export default function App() {
           )}
 
           {/* ── MOBILITY VIEW ── */}
-          {view === 'mobility' && (
-            <>
-              <div
-                style={{
-                  background: 'linear-gradient(135deg,#a78bfa15,#1e1e30)',
-                  border: '1px solid #a78bfa40',
-                  borderLeft: '3px solid #a78bfa',
-                  borderRadius: 6,
-                  padding: '13px 16px',
-                  marginBottom: 14,
-                }}
-              >
-                <div
-                  style={{
-                    fontSize: 13,
-                    fontWeight: 700,
-                    color: '#a78bfa',
-                    marginBottom: 5,
-                  }}
-                >
-                  🧘 MOBILITY — a training pillar
-                </div>
-                <div
-                  style={{ fontSize: 11, color: '#aaaacc', lineHeight: 1.6 }}
-                >
-                  Not accessory work. With the left hamstring/glute rehab,
-                  mobility is load-bearing for staying healthy enough to train.
-                  The pre-session prime especially — it's rehab activation
-                  disguised as a warm-up.
-                </div>
-              </div>
-
-              {/* Routines library */}
-              <div
-                style={{
-                  fontSize: 9,
-                  letterSpacing: 3,
-                  color: '#a78bfa',
-                  marginBottom: 8,
-                }}
-              >
-                ROUTINES · ON HAND
-              </div>
-              <div
-                style={{
-                  display: 'flex',
-                  flexDirection: 'column',
-                  gap: 8,
-                  marginBottom: 18,
-                }}
-              >
-                {MOBILITY_ROUTINES.map((r, i) => {
-                  const isOpen = mobOpen === r.id;
-                  return (
-                    <div
-                      key={r.id}
-                      style={{
-                        background: isOpen ? `${r.color}10` : '#2a2a48',
-                        border: `1px solid ${isOpen ? r.color + '50' : '#4a4a68'}`,
-                        borderLeft: `3px solid ${r.color}`,
-                        borderRadius: 6,
-                        overflow: 'hidden',
-                      }}
-                    >
-                      <div
-                        onClick={() => setMobOpen(isOpen ? null : r.id)}
-                        style={{
-                          padding: '12px 14px',
-                          cursor: 'pointer',
-                          display: 'flex',
-                          alignItems: 'center',
-                          justifyContent: 'space-between',
-                        }}
-                      >
-                        <div
-                          style={{
-                            display: 'flex',
-                            alignItems: 'center',
-                            gap: 11,
-                          }}
-                        >
-                          <span style={{ fontSize: 16 }}>{r.icon}</span>
-                          <div>
-                            <div
-                              style={{
-                                fontSize: 13,
-                                fontWeight: 700,
-                                color: '#fff',
-                              }}
-                            >
-                              {r.name}
-                            </div>
-                            <div
-                              style={{
-                                fontSize: 9,
-                                color: '#7e7e9a',
-                                marginTop: 1,
-                              }}
-                            >
-                              {r.when}
-                            </div>
-                          </div>
-                        </div>
-                        <span style={{ fontSize: 10, color: '#7e7e9a' }}>
-                          {isOpen ? '▲' : '▼'}
-                        </span>
-                      </div>
-                      {isOpen && (
-                        <div style={{ padding: '0 14px 14px' }}>
-                          <div
-                            style={{
-                              fontSize: 10,
-                              color: '#888',
-                              lineHeight: 1.6,
-                              marginBottom: 10,
-                              fontStyle: 'italic',
-                            }}
-                          >
-                            {r.why}
-                          </div>
-                          <div
-                            style={{
-                              display: 'flex',
-                              flexDirection: 'column',
-                              gap: 6,
-                              marginBottom: 10,
-                            }}
-                          >
-                            {r.blocks.map((b, j) => (
-                              <div
-                                key={j}
-                                style={{
-                                  background: '#08080d',
-                                  borderRadius: 5,
-                                  padding: '9px 11px',
-                                  borderLeft: b.rehab
-                                    ? '2px solid #34d399'
-                                    : '2px solid transparent',
-                                }}
-                              >
-                                <div
-                                  style={{
-                                    display: 'flex',
-                                    gap: 8,
-                                    alignItems: 'baseline',
-                                    marginBottom: 3,
-                                  }}
-                                >
-                                  <span
-                                    style={{
-                                      color: r.color,
-                                      flexShrink: 0,
-                                      fontSize: 11,
-                                      fontWeight: 700,
-                                    }}
-                                  >
-                                    {j + 1}
-                                  </span>
-                                  <span
-                                    style={{
-                                      fontSize: 11,
-                                      color: '#e8e8f0',
-                                      fontWeight: 600,
-                                    }}
-                                  >
-                                    {b.move}
-                                    {b.rehab && (
-                                      <span
-                                        style={{
-                                          color: '#34d399',
-                                          fontSize: 9,
-                                          fontWeight: 700,
-                                        }}
-                                      >
-                                        {' '}
-                                        · REHAB
-                                      </span>
-                                    )}
-                                  </span>
-                                </div>
-                                <div
-                                  style={{
-                                    fontSize: 10,
-                                    color: '#888',
-                                    lineHeight: 1.6,
-                                    paddingLeft: 19,
-                                  }}
-                                >
-                                  {b.visual}
-                                </div>
-                              </div>
-                            ))}
-                          </div>
-                          <div
-                            style={{
-                              background: '#08080d',
-                              borderLeft: `2px solid ${r.color}`,
-                              borderRadius: 4,
-                              padding: '9px 11px',
-                              fontSize: 10,
-                              color: '#aaaacc',
-                              lineHeight: 1.6,
-                            }}
-                          >
-                            📋 {r.note}
-                          </div>
-                        </div>
-                      )}
-                    </div>
-                  );
-                })}
-              </div>
-
-              {/* Tracking log */}
-              <div
-                style={{
-                  fontSize: 9,
-                  letterSpacing: 3,
-                  color: '#a78bfa',
-                  marginBottom: 8,
-                }}
-              >
-                RECENT · TRACKED
-              </div>
-              <div
-                style={{
-                  background: '#1e1e30',
-                  border: '1px solid #a78bfa30',
-                  borderLeft: '3px solid #a78bfa',
-                  borderRadius: 6,
-                  padding: '10px 13px',
-                  marginBottom: 10,
-                  fontSize: 10,
-                  color: '#888',
-                  lineHeight: 1.6,
-                }}
-              >
-                {MOBILITY_STREAK_NOTE}
-              </div>
-              <div style={{ display: 'flex', flexDirection: 'column', gap: 6 }}>
-                {mobilityLog.map((m, i) => {
-                  const r = MOBILITY_ROUTINES.find((x) => x.id === m.type);
-                  const col = r ? r.color : '#888';
-                  return (
-                    <div
-                      key={i}
-                      style={{
-                        background: '#2a2a48',
-                        border: '1px solid #4a4a68',
-                        borderLeft: `3px solid ${col}`,
-                        borderRadius: 6,
-                        padding: '11px 14px',
-                      }}
-                    >
-                      <div
-                        style={{
-                          display: 'flex',
-                          alignItems: 'center',
-                          justifyContent: 'space-between',
-                          marginBottom: 4,
-                        }}
-                      >
-                        <div
-                          style={{
-                            display: 'flex',
-                            alignItems: 'center',
-                            gap: 9,
-                          }}
-                        >
-                          <span style={{ fontSize: 14 }}>
-                            {r ? r.icon : '•'}
-                          </span>
-                          <div
-                            style={{
-                              fontSize: 12,
-                              fontWeight: 700,
-                              color: '#fff',
-                            }}
-                          >
-                            {m.label}
-                          </div>
-                        </div>
-                        <div style={{ fontSize: 10, color: '#7e7e9a' }}>
-                          {m.date} · {m.duration}
-                        </div>
-                      </div>
-                      <div
-                        style={{
-                          fontSize: 10,
-                          color: '#aaaacc',
-                          lineHeight: 1.5,
-                          paddingLeft: 23,
-                        }}
-                      >
-                        {m.note}
-                      </div>
-                    </div>
-                  );
-                })}
-              </div>
-              <div
-                style={{
-                  marginTop: 12,
-                  background: '#1e1e30',
-                  border: '1px solid #4a4a68',
-                  borderRadius: 6,
-                  padding: '11px 14px',
-                  fontSize: 10,
-                  color: '#7e7e9a',
-                  textAlign: 'center',
-                  lineHeight: 1.6,
-                }}
-              >
-                Report mobility sessions in chat to log them here. Pre-session
-                prime + foam roll + yoga all count.
-              </div>
-            </>
-          )}
+          {view === 'mobility' && <MobilityView />}
 
           {/* ── RECOVERY VIEW ── */}
           {view === 'recovery' && (
